Move maintenance to maintenance update command
